requirements.txt
requests
flask
flask_cors
python-dotenv